+++ portal-impl/src/com/liferay/portal/upgrade/v6_1_1/UpgradeLayout.java
@@ -85,7 +85,7 @@ public class UpgradeLayout extends UpgradeProcess {
 			rs = ps.executeQuery();
 
 			while (rs.next()) {
-				long count = rs.getLong(1);
+				int count = rs.getInt(1);
 
 				if (count > 0) {
 					return true;
